Create Hangman
The user picks a file containing a list of words and the length of the word. The list of words is read from the file and the program randomly selects a word of the appropriate length and the game begins. The player guesses a letter and either the letter is revealed in the word or the user is informed that the letter is not in the owrd. The letter is added to the list of previous guess and the user is informed if the player repeated a guess. The program determines if the player has won or lost.